Extract get_stain_vector into histomicstk
