+++ server/NucleiDetection/NucleiDetection.py
@@ -24,27 +24,6 @@ from ctk_cli import CLIArgumentParser
 import logging
 logging.basicConfig()
 
-def get_stain_vector(args, index):
-    """Get the stain corresponding to args.stain_$index and
-    args.stain_$index_vector.  If the former is not "custom", the
-    latter must be None.
-
-    """
-    args = vars(args)
-    stain = args['stain_' + str(index)]
-    stain_vector = args['stain_' + str(index) + '_vector']
-    if stain == 'custom':
-        if stain_vector is None:
-            raise ValueError('If "custom" is chosen for a stain, '
-                             'a stain vector must be provided.')
-        return stain_vector
-    else:
-        if stain_vector is None:
-            return htk_cdeconv.stain_color_map[stain]
-        raise ValueError('Unless "custom" is chosen for a stain, '
-                         'no stain vector may be provided.')
-
-
 def detect_nuclei_kofahi(im_input, args):
 
     # perform color normalization
@@ -53,7 +32,7 @@ def detect_nuclei_kofahi(im_input, args):
                                  args.reference_std_lab)
 
     # perform color decovolution
-    w = np.array([get_stain_vector(args, i) for i in 1, 2, 3]).T
+    w = np.array([htk_cdeconv.utils.get_stain_vector(args, i) for i in 1, 2, 3]).T
 
     im_stains = htk_cdeconv.color_deconvolution(im_nmzd, w).Stains
